Move UserAuth module and rename crypto helpers
