+++ src/package_manager_frontend/src/App.tsx
@@ -35,6 +35,7 @@ import ModuleCycles from './ModuleCycles';
 import { AuthProvider, getIsLocal, useAuth } from '../../lib/use-auth-client';
 import { package_manager } from '../../declarations/package_manager';
 import { cycles_ledger } from '../../declarations/cycles_ledger';
+import { getPublicKeyFromPrivateKey, signPrincipal } from "../../lib/signatures";
 
 function App() {
   return (
@@ -61,39 +62,6 @@ function App() {
   );
 }
 
-async function getPublicKeyFromPrivateKey(privateKey: CryptoKey) {
-  try {
-    // Export the private key in JWK format
-    const jwkPrivate = await crypto.subtle.exportKey("jwk", privateKey);
-
-    // Create a JWK for the public key using modulus (n) and exponent (e)
-    const jwkPublic = {
-      kty: jwkPrivate.kty, // Key type (EC)
-      crv: jwkPrivate.crv, // Curve (e.g., P-256)
-      x: jwkPrivate.x,     // X coordinate
-      y: jwkPrivate.y,     // Y coordinate
-      alg: jwkPrivate.alg, // Algorithm (e.g., ES256)
-    };
-
-    // Import the public key back into the Web Crypto API
-    const publicKey = await crypto.subtle.importKey(
-      "jwk",
-      jwkPublic,
-      {
-        name: "ECDSA",
-        namedCurve: jwkPrivate.crv, // e.g., "P-256"
-        hash: { name: "SHA-256" }, // e.g., { name: "SHA-256" }
-      },
-      true, // Extractable
-      ["verify"] // Key usages (e.g., for signing/verifying)
-    );
-
-    return publicKey;
-  } catch (error) {
-    console.error("Error deriving public key:", error);
-    throw error;
-  }
-}
 
 function GlobalUI() {
   const glob = useContext(GlobalContext);
@@ -131,16 +99,7 @@ function GlobalUI() {
           "pkcs8", glob.frontendTweakPrivKey!, {name: 'ECDSA', namedCurve: 'P-256'/*prime256v1*/}, true, ["sign"]
         );
         const pubKeyUsable = await getPublicKeyFromPrivateKey(privKeyUsable);
-        const signature = await window.crypto.subtle.sign(
-          {
-            name: 'ECDSA',
-            // namedCurve: "P-256",
-            saltLength: 32,
-            hash: "SHA-256",
-          },
-          privKeyUsable,
-          principal!.toUint8Array(),
-        );
+        const signature = await signPrincipal(privKeyUsable, principal!);
         const {spentCycles: spentBackendStr} = await bootstrapperMainIndirect.bootstrapBackend({
           frontendTweakPubKey: new Uint8Array(await window.crypto.subtle.exportKey("spki", pubKeyUsable)),
           installedModules,
